Setting AGP min version to 7.4.0
