--- agp-compatibility/agp-compatibility-7-2/build.gradle
@@ -1,13 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-java {
-    sourceCompatibility = jvmCompatibility
-    targetCompatibility = jvmCompatibility
-}
-
-dependencies {
-    implementation project(':agp-compatibility:agp-compatibility-api')
-    compileOnly "com.android.tools.build:gradle:7.2.0"
-}
