--- agp-compatibility/agp-compatibility-7-2/src/main/resources/META-INF/services/co.elastic.apm.android.agp.api.AgpCompatibilityEntrypoint
@@ -1,1 +1,0 @@
-co.elastic.apm.android.agp72.Agp72CompatibilityEntrypoint
