--- agp-compatibility/agp-compatibility-7-3/build.gradle
@@ -1,14 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-java {
-    sourceCompatibility = jvmCompatibility
-    targetCompatibility = jvmCompatibility
-}
-
-dependencies {
-    implementation project(':agp-compatibility:agp-compatibility-api')
-    implementation project(':agp-compatibility:agp-compatibility-7-2')
-    compileOnly "com.android.tools.build:gradle:7.3.0"
-}
