--- agp-compatibility/agp-compatibility-7-3/metadata/notice.properties
@@ -1,1 +1,0 @@
-dependencies.hash=D41D8CD98F00B204E9800998ECF8427E
