--- agp-compatibility/agp-compatibility-7-3/src/main/java/co/elastic/apm/android/agp73/Agp73CompatibilityEntrypoint.java
@@ -1,45 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *	http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package co.elastic.apm.android.agp73;
-
-import com.android.build.api.AndroidPluginVersion;
-
-import org.gradle.api.Project;
-
-import co.elastic.apm.android.agp.api.AgpCompatibilityEntrypoint;
-import co.elastic.apm.android.agp.api.AgpCompatibilityManager;
-import co.elastic.apm.android.agp.api.CurrentVersion;
-
-public class Agp73CompatibilityEntrypoint implements AgpCompatibilityEntrypoint {
-
-    @Override
-    public String getDescription() {
-        return "Agp compatible for 7.3+";
-    }
-
-    @Override
-    public boolean isCompatible(CurrentVersion currentVersion) {
-        return currentVersion.isEqualOrGreaterThan(new AndroidPluginVersion(7, 3));
-    }
-
-    @Override
-    public AgpCompatibilityManager provideCompatibilityManager(Project project) {
-        return new Agp73CompatibilityManager(project);
-    }
-}
