--- agp-compatibility/agp-compatibility-7-3/src/main/resources/META-INF/NOTICE
@@ -1,2 +1,0 @@
-Elastic APM Android Agent
-Copyright 2018-2022 Elasticsearch B.V.
