--- agp-compatibility/agp-compatibility-7-3/src/main/resources/META-INF/services/co.elastic.apm.android.agp.api.AgpCompatibilityEntrypoint
@@ -1,1 +1,0 @@
-co.elastic.apm.android.agp73.Agp73CompatibilityEntrypoint
