--- agp-compatibility/agp-compatibility-api/src/main/java/co/elastic/apm/android/agp/api/CurrentVersion.java
@@ -1,45 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *	http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package co.elastic.apm.android.agp.api;
-
-import com.android.build.api.AndroidPluginVersion;
-
-public class CurrentVersion {
-    private final AndroidPluginVersion current;
-
-    public CurrentVersion(AndroidPluginVersion current) {
-        this.current = current;
-    }
-
-    public boolean isEqualTo(AndroidPluginVersion other) {
-        return current.compareTo(other) == 0;
-    }
-
-    public boolean isGreaterThan(AndroidPluginVersion other) {
-        return current.compareTo(other) > 0;
-    }
-
-    public boolean isLowerThan(AndroidPluginVersion other) {
-        return current.compareTo(other) < 0;
-    }
-
-    public boolean isEqualOrGreaterThan(AndroidPluginVersion other) {
-        return isEqualTo(other) || isGreaterThan(other);
-    }
-}
