+++ android-test/build.gradle
@@ -5,7 +5,7 @@ propertiesFile.withInputStream {
 }
 
 ext {
-    jvmCompatibility = JavaVersion.VERSION_1_8
+    jvmCompatibility = JavaVersion.VERSION_11
     junit_version = '4.13.2'
     espresso_version = '3.4.0'
     mockito_version = '4.9.0'
