+++ android-test/plugin-test/build.gradle
@@ -3,8 +3,8 @@ plugins {
 }
 
 java {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
+    sourceCompatibility = jvmCompatibility
+    targetCompatibility = jvmCompatibility
 }
 
 dependencies {
