+++ android-test/plugin-test/src/test/java/co/elastic/apm/android/plugin/BaseAssetsVerificationTest.java
@@ -21,7 +21,7 @@ public abstract class BaseAssetsVerificationTest extends BaseFunctionalTest {
     }
 
     protected String getRelativePathToGeneratedAssetFile(String taskName) {
-        return "intermediates/assets/debug/" + taskName + "/" + ApmInfo.ASSET_FILE_NAME;
+        return "ASSETS/" + taskName + "/" + ApmInfo.ASSET_FILE_NAME;
     }
 
     protected Properties loadProperties(File propertiesFile) {
