+++ android-test/plugin-test/src/test/java/co/elastic/apm/android/plugin/OkHttpConfigTest.java
@@ -43,6 +43,6 @@ public class OkHttpConfigTest extends BaseAssetsVerificationTest {
 
     @Override
     protected String getAndroidGradlePluginVersion() {
-        return "7.2.0";
+        return "7.3.0";
     }
 }
