+++ settings.gradle
@@ -26,6 +26,3 @@ include ':android-sdk'
 include ':android-instrumentation'
 include ':android-common'
 include ':android-sdk-ktx'
-include ':agp-compatibility:agp-compatibility-api'
-include ':agp-compatibility:agp-compatibility-7-2'
-include ':agp-compatibility:agp-compatibility-7-3'
